random updates, remove gcc incompatability, update readme

+++ src/engine/VertexArray.hpp
@@ -25,7 +25,7 @@ public:
 		Bind();
 		vb.Bind();
 		const auto& elements = layout.GetElements();
-		unsigned int offset = 0;
+		size_t offset = 0;
 
 		for (unsigned int i = 0; i < elements.size(); i++)
 		{
